put amm jit tests in new file
